jgz80: add support for using read/write_byte() as macros
this makes it possible to use them without function call overhead,
unlike when being used as callback functions.
current compilers are unable to inline callbacks due to language
semantics, so defining a macro before including the .c file directly
(as opposed to using it as external function calls using the header)
reduces all of the overhead, which in my test improved runtime of
zexall from 2:56 to 2:00, i.e. making the core almost 33% faster.